@@ -4,32 +4,44 @@
   "private": true,
   "homepage":".",
   "dependencies": {
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
+    "@mantine/core": "^7.10.2",
+    "@mantine/hooks": "^7.10.2",
+    "@mantine/modals": "^7.10.2",
+    "@mantine/notifications": "^7.10.2",
+    "@mantine/nprogress": "^7.10.2",
+    "@testing-library/jest-dom": "^6.4.6",
+    "@testing-library/react": "^16.0.0",
+    "@testing-library/user-event": "^14.5.2",
+    "axios": "^1.7.2",
     "body-scroll-lock": "^4.0.0-beta.0",
-    "classnames": "^2.2.6",
-    "date-fns": "^2.25.0",
+    "classnames": "^2.5.1",
+    "crypto": "^1.0.1",
+    "crypto-js": "^4.2.0",
+    "date-fns": "^3.6.0",
     "draft-js": "^0.11.7",
     "lodash": "^4.17.21",
-    "react": "17.0.2",
-    "react-datepicker": "^4.3.0",
+    "moment-timezone": "^0.5.45",
+    "postcss": "^8.4.38",
+    "postcss-preset-mantine": "^1.15.0",
+    "postcss-simple-vars": "^7.0.1",
+    "react": "^18.3.1",
+    "react-datepicker": "^7.1.0",
     "react-div-100vh": "^0.7.0",
-    "react-dnd": "14.0.2",
-    "react-dnd-html5-backend": "14.0.0",
-    "react-dom": "17.0.2",
-    "react-draft-wysiwyg": "^1.14.7",
+    "react-dnd": "16.0.1",
+    "react-dnd-html5-backend": "16.0.1",
+    "react-dom": "^18.3.1",
+    "react-draft-wysiwyg": "^1.15.0",
     "react-hook-form": "^7.52.0",
     "react-loading-skeleton": "^3.4.0",
     "react-outside-click-handler": "^1.3.0",
-    "react-range": "^1.8.11",
-    "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.2",
-    "react-slick": "^0.28.1",
-    "react-tag-input": "^6.7.3",
-    "react-tooltip": "^4.2.21",
-    "recharts": "^2.1.5",
+    "react-range": "^1.8.14",
+    "react-router": "^6.23.1",
+    "react-router-dom": "^6.23.1",
+    "react-scripts": "5.0.1",
+    "react-slick": "^0.30.2",
+    "react-tag-input": "^6.10.2",
+    "react-tooltip": "^5.27.0",
+    "recharts": "^2.12.7",
     "sass": "^1.77.6",
     "use-dark-mode": "^2.3.1"
   },
@@ -56,5 +68,11 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "@babel/plugin-proposal-optional-chaining": "^7.21.0",
+    "@babel/plugin-proposal-private-methods": "^7.18.6",
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
+    "@babel/preset-env": "^7.24.7"
   }
 }
